+++ openlane/config/pdk_compat.py
@@ -214,7 +214,6 @@ def migrate_old_config(config: Mapping[str, Any]) -> Dict[str, Any]:
                 "max_ss_100C_1v60",
                 "max_ff_n40C_1v95",
             ]
-            new["SYNTH_CORNER"] = "*_ss_100C_1v60"
         elif new["PDK"].startswith("gf180mcu"):
             new["STA_CORNERS"] = [
                 "nom_tt_025C_5v00",
@@ -227,7 +226,6 @@ def migrate_old_config(config: Mapping[str, Any]) -> Dict[str, Any]:
                 "max_ss_125C_4v50",
                 "max_ff_n40C_5v50",
             ]
-            new["SYNTH_CORNER"] = "*_ss_125C_4v50"
 
         new["DEFAULT_CORNER"] = f"nom_{default_pvt}"
         new["TIMING_VIOLATION_CORNERS"] = ["*tt*"]